Instruction set almost done
Just daa to go!

Also added some graphics stuff...

+++ libGameboy/headers/cpu.h
@@ -69,7 +69,7 @@ typedef struct __GBProcessor {
         UInt8 op;
         bool prefix;
 
-        UInt8 data; // misc. data for some instructions
+        UInt16 data; // misc. data for some instructions
         SInt8 mode;
     } state;
